Add name property to Metric class, add tests for it
